# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NEXUS Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SignupInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  avatar?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserUpdate {
  name?: string;
  avatar?: string;
  currentPassword?: string;
  newPassword?: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  active: 'active',
  on_hold: 'on_hold',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface Project {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: ProjectStatus;
  createdBy: number;
  taskCount?: number;
  completedTaskCount?: number;
  memberCount?: number;
  progressPercent?: number;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export type ProjectInputStatus = typeof ProjectInputStatus[keyof typeof ProjectInputStatus];


export const ProjectInputStatus = {
  active: 'active',
  on_hold: 'on_hold',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status?: ProjectInputStatus;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  active: 'active',
  on_hold: 'on_hold',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface ProjectUpdate {
  title?: string;
  description?: string;
  status?: ProjectUpdateStatus;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  review: 'review',
  completed: 'completed',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  assignedTo?: number | null;
  assignee?: User;
  projectId: number;
  /** @nullable */
  projectTitle?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export type TaskInputStatus = typeof TaskInputStatus[keyof typeof TaskInputStatus];


export const TaskInputStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  review: 'review',
  completed: 'completed',
} as const;

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status?: TaskInputStatus;
  priority?: TaskInputPriority;
  dueDate?: string;
  assignedTo?: number;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  review: 'review',
  completed: 'completed',
} as const;

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface TaskUpdate {
  title?: string;
  description?: string;
  status?: TaskUpdateStatus;
  priority?: TaskUpdatePriority;
  dueDate?: string;
  /** @nullable */
  assignedTo?: number | null;
}

export interface TeamMember {
  id: number;
  userId: number;
  projectId: number;
  user?: User;
  joinedAt?: string;
}

export interface TeamMemberInput {
  userId: number;
}

export interface DashboardStats {
  totalTasks: number;
  completedTasks: number;
  pendingTasks: number;
  overdueTasks: number;
  totalProjects: number;
  activeProjects: number;
  completionRate: number;
  totalMembers: number;
}

export interface ActivityItem {
  id: number;
  action: string;
  userId: number;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userAvatar?: string | null;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectTitle?: string | null;
  /** @nullable */
  taskId?: number | null;
  /** @nullable */
  taskTitle?: string | null;
  timestamp: string;
}

export interface TaskBreakdown {
  projectId: number;
  projectTitle: string;
  todo: number;
  inProgress: number;
  review: number;
  completed: number;
}

export interface VelocityPoint {
  week: string;
  completed: number;
  created: number;
}